NCR Armory Update
About The Pull Request
As my first 'Hello World' project for mapping, I made an update to the NCR armory. After showing it to the NCR LM's, I got approval to add it in-server.

Why It's Good For The Game
Makes the armory less cramped and a hassle to move through. Also, a fresh coat of paint never hurts.

Changelog
🆑
change: A small map update to the NCR armory, adds no items that were not present in the old armory
/🆑